requirements.txt

   streamlit
   openai
   python-dotenv
   numpy
   pandas
   tiktoken
   llama-index
   chromadb